[Build] Work-around for new ESP8266 macro of round()
Whoever thought it would be a good idea to make round() an explicit cast to `long`....

+++ src/src/Helpers/Rules_calculate.cpp
@@ -131,7 +131,7 @@ double RulesCalculate_t::apply_unary_operator(char op, double first)
 
   switch (un_op) {
     case UnaryOperator::Not:
-      return essentiallyZero(round(first)) ? 1.0 : 0.0;
+      return essentiallyZero(roundf(first)) ? 1.0 : 0.0;
     case UnaryOperator::Log:
       return log10(first);
     case UnaryOperator::Ln:
